Got rid of mutual import
